requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
SQLAlchemy>=2.0.29,<3.0.0
PyMySQL>=1.1.0,<2.0.0
cryptography>=42.0.0,<46.0.0
PyJWT>=2.8.0,<3.0.0
passlib>=1.7.4,<2.0.0
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.2.1,<3.0.0
email-validator>=2.1.1,<3.0.0
